% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Eulerian2All}
\alias{Eulerian2All}
\title{Second-Order Eulerian Numbers}
\usage{
Eulerian2All(n)
}
\arguments{
\item{n}{A positive integer specifying the row index.}
}
\value{
A \code{bigz} vector of length \code{n} containing the sequence
\eqn{\langle\langle n, 0 \rangle\rangle, \langle\langle n, 1 \rangle\rangle, \dots, \langle\langle n, n-1 \rangle\rangle}.
}
\description{
Calculates the exact values of the second-order Eulerian numbers
\eqn{\langle\langle n, k \rangle\rangle} for a fixed row \eqn{n} and all
\eqn{k \in \{0, \dots, n-1\}}.
}
