% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Stirling2All}
\alias{Stirling2All}
\title{Stirling Numbers of the Second Kind}
\usage{
Stirling2All(n)
}
\arguments{
\item{n}{A positive integer specifying the set size.}
}
\value{
A \code{bigz} vector of length \code{n} containing the Stirling
numbers of the second kind \eqn{S(n, 1), S(n, 2), \dots, S(n, n)}.
}
\description{
Calculates the exact values of the Stirling numbers of the second kind
\eqn{S(n, k)} for a fixed \eqn{n} and all \eqn{k \in \{1, \dots, n\}}.
}
