% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStirling2-package.R
\docType{package}
\name{logStirling2-package}
\alias{logStirling2-package}
\title{logStirling2: Fast Stirling Numbers of the Second Kind}
\description{
Provides efficient tools for calculating Stirling numbers of the second kind and their logarithms. Includes an exact arbitrary-precision implementation using 'gmp' that avoids numerical cancellation, a fast C++ backend with internal caching for log-scale calculations, and Temme's asymptotic approximation for very large inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jblood94/logStirling2}
  \item Report bugs at \url{https://github.com/jblood94/logStirling2/issues}
}

}
\author{
\strong{Maintainer}: Jon Blood \email{jblood94@gmail.com}

}
\keyword{internal}
