% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stirling2direct.R
\name{stirling2direct}
\alias{stirling2direct}
\title{Stirling Numbers of the Second Kind (Exact)}
\usage{
stirling2direct(n, k)
}
\arguments{
\item{n}{Positive integer set size.}

\item{k}{Integer subset size in \code{1:n}.}
}
\value{
A \code{bigz} object.
}
\description{
Calculates the exact value of \eqn{S(n,k)} using \code{bigz} integers.
}
\details{
Implements the explicit formula for positive arguments:

\deqn{S(n,k)=\frac{1}{k!}\sum_{j=1}^k(-1)^{k-j}\binom{k}{j}j^n}
\deqn{=\frac{1}{k!}\sum_{j=1}^k\binom{-(j+1)}{k-j}j^n}

This is a "direct" calculation similar to \code{gmp::Stirling2(method =
"direct")}, but without cancellation errors for "large" n.
}
\examples{
# Basic usage
stirling2direct(5, 3)

# Comparison with the log version
mapply(\(k) log(stirling2direct(200, k)), 10:20)
logStirling2(200, 10:20)

}
\seealso{
\code{\link{logStirling2}} for log-scale calculations accepting
vectors for \code{n} and \code{k}.
}
